requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1

pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.3.0

flask==3.0.3